refactor(store-core): extract multi-client handlers into handlers/client.ts (audit P2-3)
Twelfth slice of the P2-3 split — second session-lifecycle sub-family. Move
all connected-client roster logic into a new handlers/client.ts:

- handleClientJoined / handleClientLeft / handlePrimaryChanged / handleSessionRole
- their payload interfaces (ClientJoinedResult / ClientLeftResult /
  PrimaryChanged / SessionRoleInfo)
- the VALID_DEVICE_TYPES device-type validation set
- parseConnectedClients (the auth_ok roster parser) — it shares
  VALID_DEVICE_TYPES with the roster handlers and has no internal callers,
  so it consolidates here rather than leaving a cross-file const dependency

Pure move, no logic change: both moved portions are byte-identical to the
origin/main blocks (proven by diff). Imports ConnectedClient + SessionRole
(../types). ConnectedClient and SessionRole were used only by this family,
so both are dropped from index.ts's ../types import. The barrel re-exports
via `export * from './client'`, so the public surface is unchanged.

handlers/index.ts: 3138 -> 2930 lines.
